+++ src/main/java/ru/caloricity/carbohydratesresearch/CarbohydratesResearchCreateDto.java
@@ -6,10 +6,10 @@ import jakarta.validation.constraints.NotNull;
 import java.util.UUID;
 
 public record CarbohydratesResearchCreateDto(
-        @NotNull @Min(0) Float byuksaParallelFirst,
-        @NotNull @Min(0) Float byuksaParallelSecond,
-        @NotNull @Min(0) Float byuksaAfterDryingParallelFirst,
-        @NotNull @Min(0) Float byuksaAfterDryingParallelSecond,
+        @NotNull @Min(0) Double byuksaParallelFirst,
+        @NotNull @Min(0) Double byuksaParallelSecond,
+        @NotNull @Min(0) Double byuksaAfterDryingParallelFirst,
+        @NotNull @Min(0) Double byuksaAfterDryingParallelSecond,
         @NotNull UUID probeId
 ) {
 }
