+++ src/main/java/ru/caloricity/drysubstancesresearch/DrySubstancesResearchCreateDto.java
@@ -7,10 +7,10 @@ import lombok.Builder;
 import java.util.UUID;
 
 @Builder
-public record DrySubstancesResearchCreateDto(@NotNull @Min(0) Float byuksaParallelFirst,
-                                             @NotNull @Min(0) Float byuksaParallelSecond,
-                                             @NotNull @Min(0) Float byuksaAfterDryingParallelFirst,
-                                             @NotNull @Min(0) Float byuksaAfterDryingParallelSecond,
+public record DrySubstancesResearchCreateDto(@NotNull @Min(0) Double byuksaParallelFirst,
+                                             @NotNull @Min(0) Double byuksaParallelSecond,
+                                             @NotNull @Min(0) Double byuksaAfterDryingParallelFirst,
+                                             @NotNull @Min(0) Double byuksaAfterDryingParallelSecond,
                                              @NotNull UUID probeId
 ) {
 }
