+++ src/main/java/ru/caloricity/ingredient/IngredientDto.java
@@ -7,10 +7,10 @@ import java.util.UUID;
 @Builder
 public record IngredientDto(UUID id,
                             String name,
-                            Float ediblePart,
-                            Float water,
-                            Float proteins,
-                            Float fats,
-                            Float carbohydrates
+                            Double ediblePart,
+                            Double water,
+                            Double proteins,
+                            Double fats,
+                            Double carbohydrates
 ) {
 }
