+++ src/main/java/ru/caloricity/probe/ProbeCreateDto.java
@@ -12,8 +12,8 @@ public record ProbeCreateDto(
         @NotNull @NotBlank @Size(min = 2) String name,
         @NotNull ProbeType type,
         @NotNull String code,
-        @Nullable @Min(0) Float massTheory,
-        @Nullable @Min(0) Float bankaEmptyMass,
-        @Nullable @Min(0) Float bankaWithProbeMass
+        @Nullable @Min(0) Double massTheory,
+        @Nullable @Min(0) Double bankaEmptyMass,
+        @Nullable @Min(0) Double bankaWithProbeMass
 ) {
 }
