+++ src/main/java/ru/caloricity/probeingredient/ProbeIngredientDto.java
@@ -11,7 +11,7 @@ public record ProbeIngredientDto(UUID id,
                                  LocalDateTime updatedAt,
                                  UUID probeId,
                                  UUID ingredientId,
-                                 Float gross,
-                                 Float net
+                                 Double gross,
+                                 Double net
 ) {
 }
