+++ src/main/java/ru/caloricity/probeingredient/ProbeIngredientInPageDto.java
@@ -6,12 +6,12 @@ import java.util.UUID;
 
 @Builder
 public record ProbeIngredientInPageDto(UUID id,
-                                       Float gross,
-                                       Float net,
+                                       Double gross,
+                                       Double net,
                                        String ingredientName,
-                                       Float drySubstances,
-                                       Float proteins,
-                                       Float fats,
-                                       Float carbohydrates
+                                       Double drySubstances,
+                                       Double proteins,
+                                       Double fats,
+                                       Double carbohydrates
 ) {
 }
