+++ src/main/java/ru/caloricity/probeingredient/ProbeIngredientUpdateDto.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 
 @Builder
 public record ProbeIngredientUpdateDto(
-        @Nullable @Min(0) Float gross,
-        @Nullable @Min(0) Float net
+        @Nullable @Min(0) Double gross,
+        @Nullable @Min(0) Double net
 ) {
 }
