+++ src/main/java/ru/caloricity/proteinsresearch/ProteinsResearchUpdateDto.java
@@ -3,9 +3,9 @@ package ru.caloricity.proteinsresearch;
 import jakarta.validation.constraints.NotNull;
 
 public record ProteinsResearchUpdateDto(
-        @NotNull Float titrantVolumeParallelFirst,
-        @NotNull Float titrantVolumeParallelSecond,
-        @NotNull Float controlVolume,
-        @NotNull Float coefficient
+        @NotNull Double titrantVolumeParallelFirst,
+        @NotNull Double titrantVolumeParallelSecond,
+        @NotNull Double controlVolume,
+        @NotNull Double coefficient
 ) {
 }
